APIs/app/requirements.txt
#fastapi>=0.68.0,<0.69.0
#pydantic>=1.8.0,<2.0.0
#uvicorn[standard]>=0.15.0,<0.16.0
#gunicorn>=20.1.0
crate
geojson
sqlalchemy
brotli-asgi
psycopg2
#crash
fastapi-socketio
requests
kafka-python
pandas
numpy
